% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{fit_panel_CAR_tw}
\alias{fit_panel_CAR_tw}
\title{Fit spatial Conditional Autoregressive (CAR) Model to panel data with Tweedie error distribution}
\usage{
fit_panel_CAR_tw(form, speff, B, data, fixed_speff = TRUE, sim = FALSE)
}
\arguments{
\item{form}{A formula specifying the model for the fixed effects.}

\item{speff}{A character string specifying the name of the spatial grouping variable (i.e., spatial random effect).}

\item{B}{A numeric matrix specifying the adjacency matrix (should be a matrix of 0's and 1's indicating which spatial groups are neighbors).}

\item{data}{A data frame containing the variables specified in the formula.}

\item{fixed_speff}{Logical indicating whether the spatial diffusion is fixed in time. If \code{TRUE}, this fits a model of the form
\deqn{y_{it} = x_{it}^T \beta + s_i + \rho \sum_{j\ne i}w_{ij}s_j + \epsilon_{it}}
where the spatial diffusion is constant in time.}

\item{sim}{Logical indicating whether to simulate data to be used with \pkg{DHARMa} to check quantile residuals.}
}
\value{
A fitted model object containing the estimated fixed effects, random effects, dispersion parameter, Tweedie power parameter, and simulated data (if sim = TRUE).
}
\description{
Fit spatial Conditional Autoregressive (CAR) Model to panel data with Tweedie error distribution
}
\examples{
fit_panel_CAR_tw(form = response ~ predictor1 + predictor2,
                  speff = "group",
                  B = adjacency_matrix,
                  data = my_data,
                  fixed_speff = TRUE)
}
